Added subsection describing the "diff context".
Beginning to realize that the main thing out of a file header in a diff
that matters is the identification of the target resource--the one
whose line begins with "---". Almost everything else is optional for
processing a diff later, or is interesting for human consumption at a
different point in time (e.g. to eyeball that the patch was generated
as 'diff old new' and not 'diff new old'). Also almost every tool I've
found generates something different in the 'file header' portion, yet
patch seems to be able to handle them all, I suspect because it largely
ignores everything beyond trying to figure out if the target file (---)
exists relative to its current context (current directory less -p level).